fix(lifecycle): re-claim a2a port after --force restart so instances row keeps real port
A `sac agents start --force` restart wrote the state.db `instances` row
with `a2a_port=None`, breaking `/v1/turn` routing (agent_send) even
though the sidecar bound fine and push/bus worked.

Root cause: agent_start resolves the a2a port (INSERT claim) up front,
then the --force agent_stop calls release_a2a_port (DELETE claim), so
record_local_instance read an empty a2a_ports table via get_port() and
persisted a2a_port=None.

Fix: re-resolve the port after the force-stop branches and before
record_local_instance. resolve_a2a_port is idempotent (claim_port
re-INSERTs the int already held in config.a2a.port), keeping the
a2a_ports claim table, config.a2a.port, and the instances row all
consistent.

Also splits the 716-line lifecycle.py (over the 512-line module limit)
into focused siblings (_start/_stop/_status/_runtime_select/_hook_runner/
_session_reset/_handover_loader); lifecycle.py is now a thin re-export
orchestrator so the public API and all importers are unchanged.

Regression tests (real SQLite, no mocks) exercise the force-stop ->
record path: the instances row now carries the real, non-None a2a_port
matching the live claim.
